requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
python-dotenv==1.0.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
requests==2.31.0
